requirements.txt

streamlit
llama-index
openai
python-dotenv
